add user_id to db(WIP)
