--- ene_backend/db_task.py
@@ -1,12 +1,0 @@
-import reflex as rx
-
-
-class Task(rx.Model, table=True):
-    name: str
-    priority: str
-    category: str
-    deadline: str
-    deadline_convert: str
-    hour: str
-    minute: str
-    memo: str
