+++ ene_backend/state/auth.py
@@ -1,7 +1,26 @@
+from typing import Optional
+
 import reflex as rx
 from sqlmodel import select
 
-from ene_backend.templates.template import ThemeState, User
+from ..db_model import User
+
+
+class ThemeState(rx.State):
+    """The state for the theme of the app."""
+
+    accent_color: str = "crimson"
+
+    gray_color: str = "gray"
+    user: Optional[User] = None
+
+    def check_login(self):
+        if not self.logged_in:
+            return rx.redirect("/")
+
+    @rx.var
+    def logged_in(self):
+        return self.user is not None
 
 
 class AuthState(ThemeState):
@@ -9,6 +28,9 @@ class AuthState(ThemeState):
     password: str
     confirm_password: str
     name: str | None
+
+    def get_address(self):
+        return self.address
 
     def signup(self):
         with rx.session() as session:
@@ -23,6 +45,8 @@ class AuthState(ThemeState):
             return rx.redirect("/")
 
     def login(self):
+        print(type(User.address))
+        print(type(self.address))
         with rx.session() as session:
             user = session.exec(select(User).where(User.address == self.address)).first()
             if user and user.password == self.password:
