+++ ene_backend/templates/template.py
@@ -2,13 +2,14 @@
 
 from __future__ import annotations
 
-from typing import Callable, Optional
+from typing import Callable
 
 import reflex as rx
 
 from ene_backend import styles
 from ene_backend.components.header import navi_bar
-from ene_backend.db_model import User
+
+from ..state.auth import ThemeState
 
 # Meta tags for the app.
 default_meta = [
@@ -62,23 +63,6 @@ def menu_button() -> rx.Component:
         top="2em",
         z_index="500",
     )
-
-
-class ThemeState(rx.State):
-    """The state for the theme of the app."""
-
-    accent_color: str = "crimson"
-
-    gray_color: str = "gray"
-    user: Optional[User] = None
-
-    def check_login(self):
-        if not self.logged_in:
-            return rx.redirect("/")
-
-    @rx.var
-    def logged_in(self):
-        return self.user is not None
 
 
 def template(
